+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 * Fixed require fileName for document only in base64 for send media message
 * Bug fix when sending mobile message change contact name to number in chatwoot
 * Bug fix when connecting whatsapp does not send confirmation message
+* Fixed quoted message with id or message directly
 
 ### Integrations